Remove ioctl casts that are unnecessary and wrong for 64=bit code.
git-svn-id: https://dosemu.svn.sourceforge.net/svnroot/dosemu/trunk@1513 7758b857-130e-0410-ab9c-d8735a528d18

+++ src/env/video/vc.c
@@ -421,7 +421,7 @@ set_process_control (void)
   sigaddset(&set, SIG_ACQUIRE);
   sigprocmask(SIG_UNBLOCK, &set, NULL);
 
-  if (ioctl (console_fd, VT_SETMODE, (int) &vt_mode))
+  if (ioctl (console_fd, VT_SETMODE, &vt_mode))
     v_printf ("initial VT_SETMODE failed!\n");
   v_printf ("VID: Set process control\n");
 }
@@ -432,7 +432,7 @@ clear_process_control (void)
   struct vt_mode vt_mode;
 
   vt_mode.mode = VT_AUTO;
-  ioctl (console_fd, VT_SETMODE, (int) &vt_mode);
+  ioctl (console_fd, VT_SETMODE, &vt_mode);
   setsig (SIG_RELEASE, SIG_IGN);
   setsig (SIG_ACQUIRE, SIG_IGN);
 }
